Moved prepared_wrapp_call and related to util.hpp
